refactor(BridgeMaker): 메서드 분리 addMoveCommand -> randomNumbers + moveCommand
addMoveCommand가 랜덤 숫자를 받아 'U' 혹은 'D'로 바꿔주는 역할을 할 뿐만 아니라, List<String> result에 값을 추가하는 역할까지 하고 있는 것 같아 분리하였습니다.
randomNumbers: size 만큼 랜덤 숫자 생성
moveCommand: 랜덤 숫자를 받아 'U' 혹은 'D'로 변경
makeBridge: 위 두 메서드를 이용하여 List<String>을 반환
