+++ src/main/java/bridge/domain/make_bridge/BridgeMaker.java
@@ -1,10 +1,11 @@
-package bridge.domain;
+package bridge.domain.make_bridge;
 
 import static bridge.domain.constants.MoveCommands.MOVE_DOWN_COMMAND;
 import static bridge.domain.constants.MoveCommands.MOVE_UP_COMMAND;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 /**
  * 다리의 길이를 입력 받아서 다리를 생성해주는 역할을 한다.
@@ -24,23 +25,24 @@ public class BridgeMaker {
      * @return 입력받은 길이에 해당하는 다리 모양. 위 칸이면 "U", 아래 칸이면 "D"로 표현해야 한다.
      */
     public List<String> makeBridge(final int size) {
-        List<String> results = new ArrayList<>();
-
-        for (int i = 0; i < size; i++) {
-            int randomNumber = bridgeNumberGenerator.generate();
-            addMoveCommand(results, randomNumber);
-        }
-
-        return results;
+        List<Integer> randomNumbers = randomNumbers(size);
+        return randomNumbers.stream()
+                .map(this::moveCommand)
+                .collect(Collectors.toList());
     }
 
-    private void addMoveCommand(final List<String> results, final int randomNumber) {
+    private List<Integer> randomNumbers(final int size) {
+        return IntStream.range(0, size)
+                .mapToObj(i -> bridgeNumberGenerator.generate())
+                .collect(Collectors.toList());
+    }
+
+    private String moveCommand(final int randomNumber) {
         final int NUMBER_FOR_MOVE_UP = 1;
         if (randomNumber == NUMBER_FOR_MOVE_UP) {
-            results.add(MOVE_UP_COMMAND);
-            return;
+            return MOVE_UP_COMMAND;
         }
 
-        results.add(MOVE_DOWN_COMMAND);
+        return MOVE_DOWN_COMMAND;
     }
 }
