+++ src/main/java/bridge/domain/make_bridge/BridgeNumberGenerator.java
@@ -1,4 +1,4 @@
-package bridge.domain;
+package bridge.domain.make_bridge;
 
 // TODO BridgeRandomNumberGenerator, BridgeNumberGenerator 클래스의 코드는 변경할 수 없다.
 @FunctionalInterface
