+++ src/test/java/bridge/domain/make_bridge/BridgeMakerTest.java
@@ -1,4 +1,4 @@
-package bridge.domain;
+package bridge.domain.make_bridge;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
